// src/core/physics/types.ts
export interface GlobalMacros {
  /** amount of activity */
  energy: number
  /** 0 = ORDER, 1 = CHAOS */
  chaos: number
  /** temporal scale: 0.5 = nominal, 0 = half speed, 1 = double */
  time: number
  /** spatial spread */
  space: number
}

export const DEFAULT_MACROS_GLOBAL: GlobalMacros = { energy: 0.5, chaos: 0.25, time: 0.5, space: 0.5 }